requirements.txt
bs4
requests
urllib3
selenium
tqdm
textract
xmi-reader
dkpro-cassis

certifi==2021.10.8
pdf2image==1.16.0
Pillow==8.4.0
pytesseract==0.3.8
